apps/ml/requirements.txt
# Fallback for environments without Poetry. Keep in sync with pyproject.toml.
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.0
pillow==11.0.0
ultralytics==8.3.60
torch==2.5.1
numpy==2.2.1
sentry-sdk[fastapi]==2.19.2
